Form Validation for Checkboxes & Dropdowns
Form is split up into boxes, and boxes show and hide once the previous
part of the form is filled out and validated.

Form is posted to an asp form parser at www.newmanic.com/formtest.asp

Show/Hide is broken in this build. Counter functions and if conditions
for the checkbox validation should be reviewed.